==> analysis/4-DataExploration.ipynb <==
# This notebook was used to conduct preliminary exploration of data
# As such it is not a part of the main workflow, but used just for additional insights

# Import the required libraries
import pandas as pd
import time
from datetime import timedelta, datetime
from sqlalchemy import create_engine
from sklearn.feature_extraction.text import CountVectorizer


# Create SQL engine; choose the correct path where reddit.db database file is located
engine = create_engine(f'sqlite:///C:/Users/Akshay/Desktop/ClimateChange/Database/reddit.db')


# Select all the data from Table A in the database
postsA = pd.read_sql("""select * from A order by utc asc""",engine)
postsA.head()

# Select all the data from Table B in the database
postsB = pd.read_sql("""select * from B order by utc asc""",engine)
postsB.head()

# Merge data from Table A & B
posts = pd.concat([postsA,postsB])
len(posts)

posts.tail()

# Convert the Unix epoch time to human readable dates and set index
posts['date'] = posts['utc'].map(lambda x: datetime.utcfromtimestamp(x).strftime('%Y-%m-%d'))
posts.set_index('date', inplace=True)
posts.head()

# Sort from latest to oldest
posts.sort_index(inplace=True,ascending=False)
posts.head()

# Check the number of unique users and reddit posts per subreddit/community 
users = posts[['subreddit','author','title']].groupby(['subreddit'],as_index=False).nunique()
users.sort_values(['title'],inplace=True, ascending=False)
users.reset_index(inplace=True, drop=True)
users

# Considering the "EcoInternet" subreddit to be an outlier, observed from the table above; 
# check the users(authors) and their respective post counts within "EcoInternet" community
ecoInternet = posts[posts['subreddit'] == 'EcoInternet'][['author','title']].groupby(['author'],as_index=False).nunique()
ecoInternet

# Save the data to local disk
users.to_excel('UsersCount.xlsx')
ecoInternet.to_csv('EcoInternetUsers.csv')

# Explore the "EcoInternet" subreddit outlier slightly more
ecoInternetdf = posts[posts['subreddit'] == 'EcoInternet'].copy()
ecoInternetdf.sort_index(inplace=True, ascending=False,ignore_index=True)
ecoInternetdf.tail()

# Check the daily distribution of number of posts published in "EcoInternet" mainly by 1 user over time
ecoInternetTS = ecoInternetdf[['title']].groupby(ecoInternetdf.index).nunique()
ecoInternetTS.head()

# Check the daily distribution of number of posts published by all users in all the selected subreddits over time
distribution = posts[['title']].groupby(posts.index).nunique()
distribution.head()

# Merge the two daily time-series (EcoInternet & Total Posts) in a single dataframe for further processing
distribution = distribution.merge(ecoInternetTS,how='outer',left_index=True,right_index=True)
distribution.head()

distribution.tail()

# Save the time-series on local disk for further processing
distribution.to_csv('postsTS.csv')

# Next, merge the text from "title" and "post" ie main body together to observe the overall Keywords & counts
posts['text'] = posts['title'] + "  " + posts['post']
posts.head()

# Initiate the CounteVectorizer and count the number of Keywords; mainly unigrams, bigrams and trigrams
# Rerun the cell by changing the ngram_range parameter for bigrams and trigrams once it is done running for unigrams

start = time.monotonic()

cv = CountVectorizer(ngram_range=(1,1),stop_words='english')  #Also use; ngram_range=(2,2) & (3,3) for bigrams & trigrams

cvFit = cv.fit_transform([str(posts['text'].values.tolist())])

wordList = cv.get_feature_names()
countList = cvFit.toarray().sum(axis=0)

dictionary = dict(zip(wordList,countList))
dictionary = dict(sorted(dictionary.items(), key = lambda x: x[1], reverse=True))

cvDf = pd.DataFrame(list(dictionary.items()), columns=['keywords','count'])

end = time.monotonic()

print(f"Total time taken: {timedelta(seconds=(end-start))}")

cvDf.head()

cvDf.tail()

# Finally save the "keywordsBigrams100.csv" and "keywordsTrigrams100.csv" files by repeating the same previous step
# Keywords with greater than or equal to 100 mentions only
cvDf[cvDf['count'] >= 100].to_csv("keywordsUnigrams100.csv")

# Count the number of unique Reddit users(ie authors)
posts['author'].nunique()
